% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_cleaner.R
\name{A1c_cleaner}
\alias{A1c_cleaner}
\title{A1c_cleaner() Function}
\usage{
A1c_cleaner(x)
}
\arguments{
\item{x}{Vector}
}
\description{
Clean HbA1c values.
}
\examples{
A1c_cleaner(vector)
}
\keyword{A1c_cleaner}
