% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apache2_gen.R
\name{apache2_gen}
\alias{apache2_gen}
\title{Generate APACHE II score}
\usage{
apache2_gen(data, age, bt, map, ph, pr, rr, na, kal, cr, hct, wbc, gcs, po2)
}
\arguments{
\item{data}{data.frame object}

\item{age}{Age}

\item{bt}{Body temperature}

\item{map}{Mean arterial pressure}

\item{ph}{Blood pH}

\item{pr}{Pulse rate}

\item{rr}{Respiratory rate}

\item{na}{Serum sodium}

\item{kal}{Serum potassium}

\item{cr}{Serum Creatinine}

\item{hct}{Hematocrit}

\item{wbc}{White blood cells count}

\item{gcs}{Glasgow Coma Scale}

\item{po2}{Blood pO2}
}
\description{
Generate APACHE II score
}
\keyword{apache2_gen}
