% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_base_save.R
\name{bind_base_save}
\alias{bind_base_save}
\title{Bind, extract, and save laboratory data}
\usage{
bind_base_save(
  hospital,
  ptlist_file,
  lab_dir_path,
  lab_info_file = NULL,
  clean = TRUE,
  base_window_period = 4,
  savefile_name
)
}
\arguments{
\item{hospital}{Select the hospital: seoul, bucheon, or cheonan.}

\item{ptlist_file}{Patient list file which included "id" and "orderdate" column.}

\item{lab_dir_path}{Directory path included excel files.}

\item{lab_info_file}{Path of information file.}

\item{clean}{Make a cleaned data or not.}

\item{base_window_period}{Window period for baseline laboratory data (months).}

\item{savefile_name}{The name of output file name.}
}
\description{
Bind all labs, extract baseline data, and save to CSV files.
}
\keyword{bind_base_save}
