% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_cleaner.R
\name{cl_ANA_titer}
\alias{cl_ANA_titer}
\title{cl_ANA_titer() Function}
\usage{
cl_ANA_titer(x)
}
\arguments{
\item{x}{Vector}
}
\description{
Clean ANA titer values.
}
\examples{
cl_ANA_titer(vector)
}
\keyword{cl_ANA_titer}
