% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_cleaner.R
\name{cl_extract_titer}
\alias{cl_extract_titer}
\title{cl_extract_titer() Function}
\usage{
cl_extract_titer(x)
}
\arguments{
\item{x}{Vector}
}
\description{
Extract titer values.
}
\keyword{cl_extract_titer}
