% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_cleaner.R
\name{cl_RPR_from_RPR2}
\alias{cl_RPR_from_RPR2}
\title{cl_RPR_from_RPR2() Function}
\usage{
cl_RPR_from_RPR2(x1, x2)
}
\arguments{
\item{x1}{Vector}

\item{x2}{Vector}
}
\description{
RPR cleaning from RPR2 values
}
\keyword{cl_RPR_from_RPR2}
