% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_cleaner.R
\name{cl_RUA_grade}
\alias{cl_RUA_grade}
\title{cl_RUA_grade() Function}
\usage{
cl_RUA_grade(x)
}
\arguments{
\item{x}{Vector}
}
\description{
Clean RUA results containing grade values such as 1+, 2+, 3+, and 4+.
}
\examples{
cl_RUA_grade(vector)
}
\keyword{cl_RUA_grade}
