% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_cleaner.R
\name{cl_serol_marker}
\alias{cl_serol_marker}
\title{cl_serol_marker() Function}
\usage{
cl_serol_marker(x, cutoff = 0)
}
\arguments{
\item{x}{Vector}

\item{cutoff}{cutoff value for positive}
}
\description{
Clean serologic markers values.
}
\examples{
cl_serol_marker(vector)
}
\keyword{cl_serol_marker}
