% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_cleaner.R
\name{clean_compare}
\alias{clean_compare}
\title{clean_compare() Function}
\usage{
clean_compare(data_orig, data_clean)
}
\arguments{
\item{data_orig}{a original data.frame}

\item{data_clean}{a cleaned data.frame}
}
\description{
Compare the cleaned data.frame with the original data.frame
}
\keyword{clean_compare}
