% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_function.R
\name{compare_clean}
\alias{compare_clean}
\title{compare_clean() Function}
\usage{
compare_clean(data, lab, clean_fun, ...)
}
\arguments{
\item{data}{data.frame object}

\item{lab}{variable to examine}

\item{clean_fun}{specific function cleaning a variable}

\item{...}{parameters for cleaining function (clean_fun)}
}
\description{
Compare previous values and cleaned values after applying a specific function.
}
\keyword{compare_clean}
