% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_base.R
\name{extract_base}
\alias{extract_base}
\title{Extract baseline laboratory data}
\usage{
extract_base(data, base_window_period = 4)
}
\arguments{
\item{data}{Data acquired by lab_binder function.}

\item{base_window_period}{Window period (months).}
}
\description{
Extracting baseline laboratory data.
}
\keyword{extract_base}
