% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fu.R
\name{extract_fu}
\alias{extract_fu}
\title{Extract f/u lab data}
\usage{
extract_fu(
  data,
  fu_lab_name,
  fu_intv_month = 3,
  fu_max_month = 192,
  fu_wide = TRUE
)
}
\arguments{
\item{data}{Data made by lab_binder function.}

\item{fu_lab_name}{The name of output file name.}

\item{fu_intv_month}{Specific interval (months).}

\item{fu_max_month}{Maximum month you want to see.}

\item{fu_wide}{Wide format or long format.}
}
\description{
Extracting lab data at specific intervals.
}
\keyword{extract_fu}
