% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_cleaner.R
\name{lab_cleaner}
\alias{lab_cleaner}
\title{lab_cleaner() Function}
\usage{
lab_cleaner(data, rm_empty_col = F)
}
\arguments{
\item{data}{Dataframe to be cleaned}
}
\description{
Cleaning the dataframe.
}
\keyword{lab_cleaner}
