% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_cleaner.R
\name{lab_cleaner_read}
\alias{lab_cleaner_read}
\title{lab_cleaner_read() Function}
\usage{
lab_cleaner_read(file, rm_empty_col = F)
}
\arguments{
\item{file}{file to be cleaned}
}
\description{
Cleaning the dataframe.
}
\keyword{lab_cleaner_read}
