% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lab_info}
\alias{lab_info}
\title{Information about the name of laboratory parameters.}
\format{
A data frame with 156 rows and 5 variables:
\describe{
\item{sort1}{types of samples, blood or urine}
\item{cheonan}{lab names used in cheonan}
\item{bucheon}{lab names used in bucheon}
\item{seoul}{lab names used in seoul}
\item{vname}{representative abbreviation}
}
}
\usage{
lab_info
}
\description{
A dataset containing the name of laboratory parameters and thier abbreviation.
}
\keyword{datasets}
