% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_cleaner.R
\name{numeric_string}
\alias{numeric_string}
\title{numeric_string() Function}
\usage{
numeric_string(x, exact = F, extract = F)
}
\arguments{
\item{x}{Vector}

\item{exact}{exact matching or not}

\item{extract}{string extract or detect}
}
\description{
return the result of regular expression matching of numeric string
}
\keyword{numeric_string}
