% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_function.R
\name{remove_empty_cols}
\alias{remove_empty_cols}
\title{remove_empty_cols() Function}
\usage{
remove_empty_cols(data)
}
\arguments{
\item{data}{data.frame object}
}
\description{
Show non numeric variables and their values.
}
\keyword{remove_empty_cols}
