% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_cleaner.R
\name{RUA_grade}
\alias{RUA_grade}
\title{RUA_grade() Function}
\usage{
RUA_grade(x)
}
\arguments{
\item{x}{Vector}
}
\description{
Clean RUA results containing grade values such as 1+, 2+, 3+, and 4+.
}
\examples{
RUA_grade(vector)
}
\keyword{RUA_grade}
