% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_function.R
\name{select_first_lab}
\alias{select_first_lab}
\title{select_first_lab() Function}
\usage{
select_first_lab(vector)
}
\arguments{
\item{vector}{Vector}
}
\description{
Select first lab value except NA.
}
\keyword{select_first_lab}
