% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_function.R
\name{select_non_num}
\alias{select_non_num}
\title{select_non_num() Function}
\usage{
select_non_num(data)
}
\arguments{
\item{data}{data.frame object}
}
\description{
Show non numeric variables and their values.
}
\keyword{select_non_num}
