% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_cleaner.R
\name{cl_remove_symbol}
\alias{cl_remove_symbol}
\title{cl_remove_symbol() Function}
\usage{
cl_remove_symbol(x, y)
}
\arguments{
\item{x}{Vector}
}
\description{
Remove the symbols unnecessary
}
\keyword{cl_remove_symbol}
