% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_binder.R
\name{lab_binder}
\alias{lab_binder}
\title{lab_binder() Function}
\usage{
lab_binder(
  hospital,
  ptlist_path,
  lab_dir_path,
  savefile_name,
  basic_info_path = NULL,
  clean = TRUE,
  return = FALSE
)
}
\arguments{
\item{hospital}{Select the hospital: seoul, bucheon, or cheonan.}

\item{ptlist_path}{Patient list file which included "id" and "orderdate" column.}

\item{lab_dir_path}{Directory path included excel files.}

\item{savefile_name}{The name of output file name.}

\item{basic_info_path}{Path of information file.}

\item{clean}{Make a cleaned data or not.}
}
\description{
This function tie multiple excel files containing laboratory data to one file.
}
\examples{
lab_binder(
  hospital = "seoul",
  ptlist_path = "lab/seoul_remain.xlsx",
  lab_dir_path = "lab/seoul_remain",
  savefile_name = lab_dir_path,
)
}
\keyword{lab_binder}
