% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_cleaner.R
\name{lab_cleaner}
\alias{lab_cleaner}
\title{lab_cleaner() Function}
\usage{
lab_cleaner(data)
}
\arguments{
\item{x}{Dataframe}
}
\description{
Cleaning of dataframe.
}
\examples{
lab_cleaner(data)
}
\keyword{lab_cleaner}
